# Reuse `dpnp.nan_to_num` in `dpnp.nansum` and `dpnp.nanprod`