This PR proposes the use of `nan_to_num` over `_replace_nan` in `nansum`, `nanprod`, `nancumsum`, and `nancumprod` using new internal function `_replace_nan_no_mask`.

- [X] Have you provided a meaningful PR description?
- [ ] Have you added a test, reproducer or referred to issue with a reproducer?
- [ ] Have you tested your changes locally for CPU and GPU devices?
- [ ] Have you made sure that new changes do not introduce compiler warnings?
- [ ] Have you checked performance impact of proposed changes?
- [ ] If this PR is a work in progress, are you filing the PR as a draft?