@ndgrigorian 2025-03-01
![Screenshot 2025-02-28 193645](https://github.com/user-attachments/assets/bedacc53-02ea-4cda-a1b0-dc0331272d29)

The warning rendered quite nicely. I recommend we implement such a warning everywhere that a function synchronizes in some future PR.

`dpctl` may need to do this as well for, i.e., boolean indexing, repeat